refactor: moved classes and socket to services directory
