+++ src/services/socket.ts
@@ -1,5 +1,6 @@
 import { io } from 'socket.io-client';
 
-const URL = 'https://signal.noop.live';
+// const URL = 'https://signal.noop.live';
+const URL = 'http://localhost:3000';
 
 export const socket = io(URL, { autoConnect: false });
